Refactor rollbacks (#197)
* Create World struct

* Add storage_read/write for World

* Add WorldSnapshot to frame struct

* Add rollbacks on ret and inexplicit panics

* Adapt far_call and near_call to new world struct

* Adapt log and decommit opcodes to world

* Adapt vm to world

* Fix transient_stroage_read on world

* Add event to world and made it rollbackable

* Remove unnecessary results in World impl

* Add getters for world rollbackable fields

* Re-add InMemory storage

* Temporarily update submodule to point to adaptation branch

* Fix submodule branch

* Update makefile to update submdodule to specified branch

* Remove submodule update remote

* Add missing rollback on initial_frames

* Move rollback functions to its own file

* Hide borrows

* Refactor struct names

Now the VmState struct refers to the World and the VmState before got renamed for Execution.

* Cargo fmt

* Add update remote in submodule

* Remove submodule update remote

* Update submodule

* Remove branch from .gitmodule

* Add EOL to .gitmodukes [skip-ci]
