+++ src/op_handlers/precompile_call.rs
@@ -9,16 +9,16 @@ use zkevm_opcode_defs::{
 use crate::{
     address_operands::{address_operands_read, address_operands_store},
     eravm_error::EraVmError,
+    execution::Execution,
     precompiles::{
         ecrecover::ecrecover_function, keccak256::keccak256_rounds_function,
         secp256r1_verify::secp256r1_verify_function, sha256::sha256_rounds_function,
     },
-    state::VMState,
     value::TaggedValue,
     Opcode,
 };
 
-pub fn precompile_call(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn precompile_call(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, src1) = address_operands_read(vm, opcode)?;
     let aux_data = PrecompileAuxData::from_u256(src1.value);
