+++ src/op_handlers/ret.rs
@@ -3,8 +3,9 @@ use zkevm_opcode_defs::RetOpcode;
 
 use crate::{
     eravm_error::EraVmError,
+    execution::Execution,
+    rollbacks::Rollbackable,
     state::VMState,
-    store::StateStorage,
     value::{FatPointer, TaggedValue},
     Opcode,
 };
@@ -16,7 +17,7 @@ fn is_failure(return_type: RetOpcode) -> bool {
 }
 
 fn get_result(
-    vm: &mut VMState,
+    vm: &mut Execution,
     reg_index: u8,
     return_type: RetOpcode,
 ) -> Result<TaggedValue, EraVmError> {
@@ -32,10 +33,9 @@ fn get_result(
 }
 
 pub fn ret(
-    vm: &mut VMState,
+    vm: &mut Execution,
     opcode: &Opcode,
-    state_storage: &mut StateStorage,
-    transient_storage: &mut StateStorage,
+    state: &mut VMState,
     return_type: RetOpcode,
 ) -> Result<bool, EraVmError> {
     let is_failure = is_failure(return_type);
@@ -44,17 +44,13 @@ pub fn ret(
     vm.flag_lt_of = return_type == RetOpcode::Panic;
     vm.flag_gt = false;
 
-    if is_failure {
-        state_storage.rollback(&vm.current_frame()?.storage_snapshot);
-        transient_storage.rollback(&vm.current_frame()?.transient_storage_snapshot);
-    }
-
     if vm.in_near_call()? {
         let previous_frame = vm.pop_frame()?;
         if opcode.flag0_set {
             let to_label = opcode.imm0;
             vm.current_frame_mut()?.pc = to_label as u64;
         } else if is_failure {
+            state.rollback(previous_frame.snapshot);
             vm.current_frame_mut()?.pc = previous_frame.exception_handler;
         } else {
             vm.current_frame_mut()?.pc += 1;
@@ -69,6 +65,7 @@ pub fn ret(
         let previous_frame = vm.pop_frame()?;
         vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
         if is_failure {
+            state.rollback(previous_frame.snapshot);
             vm.current_frame_mut()?.pc = previous_frame.exception_handler;
         } else {
             vm.current_frame_mut()?.pc += 1;
@@ -76,6 +73,9 @@ pub fn ret(
 
         Ok(false)
     } else {
+        if is_failure {
+            state.rollback(vm.current_frame()?.snapshot.clone());
+        }
         if return_type == RetOpcode::Panic {
             return Ok(true);
         }
@@ -85,22 +85,16 @@ pub fn ret(
     }
 }
 
-pub fn inexplicit_panic(
-    vm: &mut VMState,
-    state_storage: &mut StateStorage,
-    transient_storage: &mut StateStorage,
-) -> Result<bool, EraVmError> {
+pub fn inexplicit_panic(vm: &mut Execution, state: &mut VMState) -> Result<bool, EraVmError> {
     vm.flag_eq = false;
     vm.flag_lt_of = true;
     vm.flag_gt = false;
-
-    state_storage.rollback(&vm.current_frame()?.storage_snapshot);
-    transient_storage.rollback(&vm.current_frame()?.transient_storage_snapshot);
 
     if vm.in_near_call()? {
         let previous_frame = vm.pop_frame()?;
         vm.current_frame_mut()?.pc = previous_frame.exception_handler;
         vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
+        state.rollback(previous_frame.snapshot);
 
         Ok(false)
     } else if vm.in_far_call() {
@@ -111,15 +105,17 @@ pub fn inexplicit_panic(
         let previous_frame = vm.pop_frame()?;
         vm.current_frame_mut()?.gas_left += previous_frame.gas_left;
         vm.current_frame_mut()?.pc = previous_frame.exception_handler;
+        state.rollback(previous_frame.snapshot);
         Ok(false)
     } else {
+        state.rollback(vm.current_frame()?.snapshot.clone());
         Ok(true)
     }
 }
 
 // When executing a far_call, if the opcode fails, we need to run the exception handler provided in the args
 // We don't need to: run ret.panic, pop a frame and run its exception handler
-pub fn panic_from_far_call(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn panic_from_far_call(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let far_call_exception_handler = opcode.imm0 as u64;
     let result = TaggedValue::new_pointer(U256::zero());
     vm.register_context_u128 = 0_u128;
