+++ src/op_handlers/unimplemented.rs
@@ -1,7 +1,7 @@
 use crate::eravm_error::{EraVmError, OpcodeError};
-use crate::{opcode::Opcode, state::VMState};
+use crate::{execution::Execution, opcode::Opcode};
 
-pub fn unimplemented(_vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn unimplemented(_vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     eprintln!("Unimplemented instruction: {:?}!", opcode.variant);
     Err(OpcodeError::UnimplementedOpcode.into())
 }
