+++ src/tracers/last_state_saver_tracer.rs
@@ -1,5 +1,5 @@
 use crate::eravm_error::EraVmError;
-use crate::{state::VMState, Opcode};
+use crate::{execution::Execution, Opcode};
 
 use super::tracer::Tracer;
 use u256::H160;
@@ -9,13 +9,13 @@ use zkevm_opcode_defs::RetOpcode;
 
 #[derive(Debug)]
 pub struct LastStateSaverTracer {
-    pub vm_state: VMState,
+    pub vm_state: Execution,
 }
 
 impl LastStateSaverTracer {
     pub fn new() -> Self {
         Self {
-            vm_state: VMState::new(
+            vm_state: Execution::new(
                 vec![],
                 vec![],
                 H160::zero(),
@@ -37,7 +37,7 @@ impl Default for LastStateSaverTracer {
 }
 
 impl Tracer for LastStateSaverTracer {
-    fn before_execution(&mut self, opcode: &Opcode, vm: &mut VMState) -> Result<(), EraVmError> {
+    fn before_execution(&mut self, opcode: &Opcode, vm: &mut Execution) -> Result<(), EraVmError> {
         if opcode.variant == Variant::Ret(RetOpcode::Ok) {
             self.vm_state = vm.clone();
         }
